code/Django/requirements.txt
Django==4.2.2
django-browser-reload==1.10.0
django-cors-headers==4.0.0
django-filter==23.2
django-tailwindcss==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
